Functions

Functions

Working with voxel images

FinEtoolsVoxelMesher.VoxelBoxModule.vtkexport — Method.

vtkexport(theFile::String, V::VoxelBoxVolume{CoordT,DataT}) where {CoordT<:Number,DataT<:Number}

Compute.

source

FinEtoolsVoxelMesher.VoxelBoxModule.fillsolid! — Method.

fillsolid!(V::VoxelBoxVolume,
    f::SolidCF, fillvalue::DataT) where {DataT<:Number}

Filled a solid using a solid characteristic function.

source

FinEtoolsVoxelMesher.VoxelBoxModule.fillvolume! — Method.

fillvolume!(V::VoxelBoxVolume, fillvalue::DataT) where {DataT<:Number}

Fill the volume with a given value.

source

FinEtoolsVoxelMesher.VoxelBoxModule.pad — Method.

pad(V::VoxelBoxVolume, ipad, jpad, kpad, padvalue)

Pad voxel box with a constant value.

source

FinEtoolsVoxelMesher.VoxelBoxModule.solidbox — Method.

solidbox(corner1::Tuple{CoordT, CoordT, CoordT},
    corner2::Tuple{CoordT, CoordT, CoordT}) where {CoordT<:Number}

Solid box with faces aligned with the global Cartesian axes.

source

FinEtoolsVoxelMesher.VoxelBoxModule.solidhalfspace — Method.

solidhalfspace(center::Tuple{CoordT, CoordT, CoordT},
    normal::Tuple{CoordT, CoordT, CoordT}) where {CoordT<:Number}

Solid halfspace.

source

FinEtoolsVoxelMesher.VoxelBoxModule.solidsphere — Method.

solidsphere(center::Tuple{CoordT, CoordT, CoordT}, r::CoordT) where {CoordT<:Number}

Solid sphere.

source

FinEtoolsVoxelMesher.VoxelBoxModule.threshold — Method.

threshold(V, threshold_value, voxel_below, voxel_above)

Threshold the data.

source

FinEtoolsVoxelMesher.VoxelBoxModule.trim — Method.

trim(V::VoxelBoxVolume, emptyvalue)

Trim off pieces of the volume that consist only of the empty value.

source

Meshing and remeshing

FinEtoolsVoxelMesher.VoxelTetMeshingModule.mesh! — Function.

mesh!(self::ImageMesher, stretch::FFlt = 1.2)

Perform a meshing step.

If no mesh exists, the initial mesh is created; otherwise a coarsening sequence of coarsen surface -> smooth -> coarsen volume -> smooth is performed.

After meshing the vertices, tetrahedra, and material identifiers, can be retrieved as self.v, self.t, and self.tmid.

source

FinEtoolsVoxelMesher.VoxelTetMeshingModule.volumes — Method.

volumes(self::ImageMesher)

Compute tetrahedral volumes in the current mesh.

source

FinEtoolsVoxelMesher.TetRemeshingModule.coarsen — Method.

coarsen(t::Array{Int, 2}, v::Array{Float64, 2}, tmid::Vector{Int}, options)

Coarsen a T4 (tetrahedral) mesh.

Arguments

t = array with vertex numbers, one per tetrahedron in each row v= array of vertex coordinates, one per vertex in each row tmid = tetrahedron material identifiers, one for tetrahedron in each row options = structure with optional fields bv=array for all vertices in the input array v. Is the vertex on the boundary? True or false. The boundary vertices are in layer 1. The vertex layer numbers increase away from the boundary, and the bigger the vertex layer number the bigger the mesh size. desiredts=mesh size desired for vertices in layer 1, here mesh size is the length of an edge stretch=the mesh size should increase by this much within one layer of elements, default is 1.25 nblayer=number of boundary layers where the mesh size should not increase, default is 1 surfacecoarsening = is the coarsening intended for the interior or for the surface? default is false, which means the default is coarsening of the interior. preservethin= when coarsening the surface, should features which are thin be unaffected by the coarsening? Here thin means consisting of only "surface" vertices. vertexweight= weight of vertices, one per vertex; weight <= 1.0 is ignored, but weight>1.0 is used to "stretch" edges connected to the vertex. In this way one may preserve certain edges by assigning larger weight to their vertices. Default is vertex_weight=

Output

t,v,tmid = new arrays

source