Types

Types

Working with voxel images

Meshing and remeshing

FinEtoolsVoxelMesher.VoxelTetMeshingModule.ImageMesher — Type.

ImageMesher{CoordT,DataT}

Tetrahedral image mesher.

source

FinEtoolsVoxelMesher.TetRemeshingModule._IntegerBuffer — Type.

_IntegerBuffer

This data structure is introduced in order to avoid constant allocation/deallocation of integer arrays. The buffer is allocated once and then reused millions of times. The memory is never released. The buffer needs to be accessed through the methods below, not directly.

source